condense examples
