--- .storybook/stories/DataFiltering/ListFilteringLargeExample.tsx
@@ -1,50 +1,0 @@
-import { ListFilteringLarge } from './ListFilteringLarge';
-import { useSortAndFilter } from './lib/useSortAndFilter';
-import { generateTableCaption, useData } from './lib/utils';
-
-export const ListFilteringLargeExample = (args: { showsTags: boolean }) => {
-	// This is for demo purposes only, and will be deleted after the next review.
-	const { showsTags } = args;
-
-	const {
-		filters,
-		pagination,
-		resetFilters,
-		removeFilter,
-		setFilters,
-		setPagination,
-		setSort,
-		sort,
-	} = useSortAndFilter({});
-
-	const { loading, data, totalPages, totalItems } = useData({
-		filters,
-		pagination,
-		sort,
-	});
-
-	const tableCaption = generateTableCaption({
-		loading,
-		totalItems,
-		pagination,
-	});
-
-	return (
-		<ListFilteringLarge
-			data={data}
-			filters={filters}
-			loading={loading}
-			pagination={pagination}
-			resetFilters={resetFilters}
-			setFilters={setFilters}
-			removeFilter={removeFilter}
-			setPagination={setPagination}
-			setSort={setSort}
-			sort={sort}
-			tableCaption={tableCaption}
-			totalPages={totalPages}
-			totalItems={totalItems}
-			showsTags={showsTags}
-		/>
-	);
-};
